rename files to match prior object renaming
