+++ packages/pipeline/src/entities/index.ts
@@ -7,7 +7,7 @@ export { DexTrade } from './dex_trade';
 export { ExchangeCancelEvent } from './exchange_cancel_event';
 export { ExchangeCancelUpToEvent } from './exchange_cancel_up_to_event';
 export { ExchangeFillEvent } from './exchange_fill_event';
-export { NonfungibleDotComTrade } from './nft_trade';
+export { NonfungibleDotComTrade } from './nonfungible_dot_com_trade';
 export { OHLCVExternal } from './ohlcv_external';
 export { Relayer } from './relayer';
 export { SraOrder } from './sra_order';
