+++ packages/pipeline/src/scripts/pull_nonfungible_dot_com_trades.ts
